+++ polaris-migrator/src/index.ts
@@ -1,7 +1,5 @@
-export {run} from './cli';
-
-export {cliConfig} from './constants';
-export type {CLIConfig} from './constants';
+export {run, cliConfig} from './cli';
+export type {CLIConfig} from './cli';
 
 export {migrate} from './migrate';
 export type {MigrateOptions} from './migrate';