+++ polaris-migrator/src/migrations/react-rename-component-prop/tests/react-rename-component-prop-with-new-value.output.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+declare function MyComponent(props: any): JSX.Element;
+declare function Child(props: any): JSX.Element;
+
+export function App() {
+  return (
+    <MyComponent newProp="new-value">
+      Hello
+      <Child prop="value" />
+    </MyComponent>
+  );
+}